requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
black==24.4.0
Brotli==1.1.0
click==8.1.7
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
django-graphene-auth==1.1.1
django-graphql-jwt==0.4.0
graphene==3.3
graphene-django==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==21.2.0
h11==0.14.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
promise==2.3
psycopg==3.1.18
psycopg2-binary==2.9.9
PyJWT==2.8.0
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.11.0
uvicorn==0.29.0
whitenoise==6.6.0
